+++ tests/testthat/test-grouped_summary.R
@@ -3,6 +3,7 @@
 testthat::test_that(
   desc = "grouped_summary with numeric measures",
   code = {
+    testthat::skip_on_cran()
     testthat::skip_if(getRversion() < "3.6")
     set.seed(123)
 
@@ -71,6 +72,7 @@ testthat::test_that(
 testthat::test_that(
   desc = "grouped_summary with factor measures",
   code = {
+    testthat::skip_on_cran()
     testthat::skip_if(getRversion() < "3.6")
     set.seed(123)